Extracted target dir creation
